+++ test/twg2/logging/test/LogMultiTest.java
@@ -4,12 +4,12 @@ import java.util.logging.Level;
 
 import org.junit.Test;
 
-import twg2.logging.LogWrapperImpl;
-import twg2.logging.LogWrapperMulti;
-import twg2.logging.Logging;
-import twg2.logging.LoggingImpl;
-import twg2.logging.LoggingMulti;
-import twg2.logging.LoggingPrefixFormat;
+import twg2.logging.LoggerImpl;
+import twg2.logging.LoggerMulti;
+import twg2.logging.LogService;
+import twg2.logging.LogServiceImpl;
+import twg2.logging.LogServiceMulti;
+import twg2.logging.LogPrefixFormat;
 
 /**
  * @author TeamworkGuy2
@@ -20,11 +20,11 @@ public class LogMultiTest {
 	@Test
 	public void loggingMultiTest() {
 		String pre = "logging - ";
-		LoggingImpl log1 = new LoggingImpl(Level.FINE, System.out, LoggingPrefixFormat.LEVEL_AND_CLASS);
-		LoggingImpl log2 = new LoggingImpl(Level.INFO, System.out, LoggingPrefixFormat.LEVEL_AND_CLASS);
+		LogServiceImpl log1 = new LogServiceImpl(Level.FINE, System.out, LogPrefixFormat.LEVEL_AND_CLASS);
+		LogServiceImpl log2 = new LogServiceImpl(Level.INFO, System.out, LogPrefixFormat.LEVEL_AND_CLASS);
 
 		@SuppressWarnings("resource")
-		LoggingMulti logMulti = new LoggingMulti(new Logging[] { log1, log2 });
+		LogServiceMulti logMulti = new LogServiceMulti(new LogService[] { log1, log2 });
 
 		logMulti.log(Level.FINER, LogMultiTest.class, "!! " + pre + "shouldn't see");
 		logMulti.log(Level.CONFIG, LogMultiTest.class, pre + "should log once");
@@ -35,14 +35,14 @@ public class LogMultiTest {
 	@Test
 	public void logWrapperMultiTest() {
 		String pre = "wrapped - ";
-		LoggingImpl logging1 = new LoggingImpl(Level.FINE, System.out, LoggingPrefixFormat.LEVEL_AND_CLASS);
-		LoggingImpl logging2 = new LoggingImpl(Level.INFO, System.out, LoggingPrefixFormat.LEVEL_AND_CLASS);
+		LogServiceImpl logging1 = new LogServiceImpl(Level.FINE, System.out, LogPrefixFormat.LEVEL_AND_CLASS);
+		LogServiceImpl logging2 = new LogServiceImpl(Level.INFO, System.out, LogPrefixFormat.LEVEL_AND_CLASS);
 
-		LogWrapperImpl log1 = new LogWrapperImpl(logging1, String.class);
-		LogWrapperImpl log2 = new LogWrapperImpl(logging2, Number.class);
+		LoggerImpl log1 = new LoggerImpl(logging1, String.class);
+		LoggerImpl log2 = new LoggerImpl(logging2, Number.class);
 
 		@SuppressWarnings("resource")
-		LogWrapperMulti logMulti = new LogWrapperMulti(new LogWrapperImpl[] { log1, log2 });
+		LoggerMulti logMulti = new LoggerMulti(new LoggerImpl[] { log1, log2 });
 
 		logMulti.log(Level.FINER, "!! " + pre + "shouldn't see");
 		logMulti.log(Level.CONFIG, pre + "should log once");